作成シート内でスキャンオプションへ遷移するよう統合
- ScanCapturePanel: スキャンオプション+撮影フロー(抽出オプション、
  英検級、カメラ/ライブラリ、複数枚トレイ、処理オーバーレイ)を
  ScanCaptureModalから抽出し再利用可能に。トレイとオーバーレイは
  ポータルで全画面表示
- CreateWordbookSheet: 「写真でスキャン」選択時に別モーダルへ
  切り替えず、同じシート内でスキャンオプションステップに遷移
  (戻るボタン付き)。UIをScanCaptureModalのパターン(インク反転の
  選択行+右ラジオ、プレートCTA)に統一
- 「カメラで撮影」押下時はカメラを即起動せず、0枚状態のトレイ
  (最初の1枚を促すページ)を表示
- 撮影トレイのシャッター左にあった重ねスタックプレビューを削除
- bottom-nav: ScanCaptureModalの直接起動を撤去しCreateWordbookSheetに一本化
